refactor: 예약 확정 함수 리팩토링
