+++ src/main/kotlin/com/reservation/ticket/concert/application/service/ReservationService.kt
@@ -20,12 +20,6 @@ import java.util.UUID
 class ReservationService(
     private val reservationRepository: ReservationRepository,
     private val seatRepository: SeatRepository,
-    private val userRepository: UserRepository,
-    private val concertRepository: ConcertRepository,
-    private val pointRepository: PointRepository,
-    private val paymentRepository: PaymentRepository,
-    private val queueRepository: QueueRepository,
-    private val userService: UserService,
 ) {
 
     fun get(id: Long): Reservation {
@@ -38,77 +32,9 @@ class ReservationService(
         return reservationRepository.save(reservation)
     }
 
-    // 결제 및 예약 상태 변경
-    @Transactional
-    fun confirmReservation(reservationId: Long): String {
-
-        val reservation = reservationRepository.findById(reservationId).orElseThrow {
-            throw IllegalArgumentException("해당 예약이 존재하지 않습니다.")
-        }
-
-        // 비관락이 들어감
-        val user = userRepository.findById(reservation.userId).orElseThrow {
-            throw IllegalArgumentException("해당 유저를 찾을 수 없습니다.")
-        }
-
-        concertRepository.findById(reservation.concert.id).orElseThrow {
-            throw IllegalArgumentException("콘서트가 존재하지 않습니다.")
-        }
-
-        val seat = seatRepository.findWriteLockById(reservation.seat.id).orElseThrow{
-            throw IllegalArgumentException("존재하지 않은 좌석 입니다.")
-        }
-
-        val queue = queueRepository.findByUserId(user.id)
-            ?: throw IllegalArgumentException("존재하지 않는 대기열 토큰 입니다.")
-
-        if (reservation.createdAt?.plusMinutes(5)!!.isAfter(LocalDateTime.now())) {
-            //대기열 삭제
-            queueRepository.delete(queue)
-            // 좌석 예약 가능으로 변경
-            seat.isAvailable = true
-            seatRepository.save(seat)
-            // 예약 취소처리
-            reservation.status = ReservationStatus.CANCELLED
-            reservationRepository.save(reservation)
-            throw IllegalArgumentException("예약 가능 시간이 만료되었습니다.")
-        }
-
-
-        // 상태를 결제 완료로 변경
-        if (reservation.status != ReservationStatus.RESERVED) {
-            throw IllegalArgumentException("결제 완료 처리를 할 수 없는 상태입니다.")
-        }
-
-        val point = pointRepository.findByUserId(user.id)
-            ?: throw IllegalArgumentException("포인트가 설정되어 있지 않습니다.")
-        // 포인트 잔액이 좌석 금액보다 적으면 예외 처리
-        if (point.amount < reservation.seat.price) {
-            throw IllegalArgumentException("포인트 잔액이 부족합니다.")
-        }
-
-        // 포인트 차감
-        point.amount -= reservation.seat.price
-        pointRepository.save(point)
-
-        // 결제 정보 저장
-        val payment = Payment(
-            userId = reservation.userId,
-            reservationId = reservation.id,
-            amount = point.amount,
-            type = PaymentType.SPEND
-        )
-        // 히스토리 저장
-        paymentRepository.save(payment)
-
-        // 예약 정보 저장
+    fun confirm(reservation: Reservation): Reservation {
         reservation.status = ReservationStatus.CONFIRMED
-        reservationRepository.save(reservation)
-
-        //대기열 삭제
-        queueRepository.delete(queue)
-
-        return "예약이 결제 완료 처리되었습니다."
+        return reservationRepository.save(reservation)
     }
 
     // 예약 취소
